JKI: MarketData features removed

+++ masterDataManager/DataLayer/Models/EvaluationTick.cs
@@ -8,7 +8,7 @@ namespace DataLayer.Models
     public class EvaluationTick : IdEntity
     {
         public DateTime TimeStamp { get; set; }
-        public double BtcValue { get; set; }
-        public double UsdValue { get; set; }
+        public decimal BtcValue { get; set; }
+        public decimal UsdValue { get; set; }
     }
 }
